+++ apps/web/src/app/(default)/content.tsx
@@ -8,7 +8,8 @@ import type { Registration } from "@echo-webkom/db/schemas";
 import { Container } from "@/components/container";
 import { JobAdPreview } from "@/components/job-ad-preview";
 import { PostPreview } from "@/components/post-preview";
-import { getRegistrationsByHappeningId, getSpotRangeByHappeningId } from "@/data/spotrange/queries";
+import { getRegistrationsByHappeningId } from "@/data/registrations/queries";
+import { getSpotRangeByHappeningId } from "@/data/spotrange/queries";
 import { fetchHomeHappenings } from "@/sanity/happening/requests";
 import { fetchAvailableJobAds } from "@/sanity/job-ad";
 import { fetchPosts } from "@/sanity/posts/requests";
