+++ apps/web/src/data/registrations/cache-keys.ts
@@ -1,0 +1,3 @@
+export const cacheKeyFactory = {
+  registrations: (id: string) => `registration-${id}`,
+};
