+++ apps/web/src/data/spotrange/cache-keys.ts
@@ -1,4 +1,3 @@
 export const cacheKeyFactory = {
-  spots: "posts",
-  registrations: (id: string) => `registration-${id}`,
+  spots: "spots",
 };
