+++ apps/web/tsconfig.json
@@ -21,6 +21,7 @@
     "**/*.cjs",
     "**/*.mjs",
     ".next/types/**/*.ts",
+    "src/data/registrations",
   ],
   "exclude": ["node_modules"],
 }
